+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+  * Fixed handling of configurable type aliases
+
 ### 0.2.0
   * Refactored translation units visitors from libclang to Clang LibTooling (#50)
   * Fixed root namespace handling (#45)